requirements.txt
django-sass-processor==0.4.6
libsass==0.13.4
Pillow==5.0.0
psycopg2==2.7.3.2
pyBarcode==0.7
pytz==2017.3
reportlab==3.4.0
six==1.11.0
